% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_revniron_to_gitignore.R
\name{add_revniron_to_gitignore}
\alias{add_revniron_to_gitignore}
\title{Adds.}
\usage{
add_revniron_to_gitignore()
}
\value{
\code{add_revniron_to_gitignore}: \code{[NULL]}
}
\description{
Renviron to .gitignore
}
