% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snaphot_test.R
\name{add_snapshot_test}
\alias{add_snapshot_test}
\title{add a new snapshots to an existing test.}
\usage{
add_snapshot_test(name, code)
}
\arguments{
\item{name}{\code{[string]} the test name}

\item{code}{\code{[expr]} the code to test, wrapped in brackets.}
}
\value{
\code{add_snapshot_test}: \code{[invisible(NULL)]} Writes a test file to the tests/thatthat directory
}
\description{
Any line without an assignment will be wrapped in \code{expect_snapshot()}.
}
