% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_reactive.R
\name{assert_reactive}
\alias{assert_reactive}
\title{check if an object is a reactive expression.}
\usage{
assert_reactive(id, x, type = NULL, ...)
}
\arguments{
\item{id}{\code{[string]}}

\item{type}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{assert_reactive}: TRUE if valid, message if invalid
}
\description{
check if an object is a reactive expression.
}
\examples{
 if (interactive()) {
 mod_ui <- function(id) {
 ns <- NS(id)
 tagList()
 }
 mod_server <- function(id, x) {
 moduleServer(id, function(input, output, session) {
 ns <- session$ns
 x <- assert_reactive(id, x, type = 'character')
 observe({
 print(x())
 })
 })
 }
 library(shiny)
 ui <- fluidPage(tags$h1('mod'), )
 server <- function(input, output, session) {
 t <- reactive({
 1
 })
 mod_server('test', x = t)
 }
 shinyApp(ui, server)
 }
}
