% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_sdevtools_proj.R
\name{assert_sdevtools_proj}
\alias{assert_sdevtools_proj}
\title{Assert that the current dir is an sDevTools Project.}
\usage{
assert_sdevtools_proj(dir = getwd())
}
\arguments{
\item{dir}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{assert_sdevtools_proj}: Throw an error if FALSE
}
\description{
Assert that the current dir is an sDevTools Project.
}
