% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_reload_gh.R
\name{build_and_reload_gh}
\alias{build_and_reload_gh}
\title{build a package from github.}
\usage{
build_and_reload_gh(
  package = current_pkg(),
  repo = "theorschrock1",
  auth_token = git_token(TRUE)
)
}
\arguments{
\item{package}{\code{[package]}  Defaults to \code{current_pkg()}}

\item{repo}{\code{[string]}  Defaults to \code{'theorschrock1'}}

\item{auth_token}{\code{[string]}  All non-missing elements must comply to regex pattern \code{'''^ghp'''}.}
}
\value{
\code{build_and_reload_gh}: \code{[invisible(NULL)]}
}
\description{
build a package from github.
}
