% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dirs_for_sdev.R
\name{build_dirs_for_sdev}
\alias{build_dirs_for_sdev}
\title{Build the directories used by sDevTools.}
\usage{
build_dirs_for_sdev()
}
\value{
\code{build_dirs_for_sdev}: NULL
}
\description{
Build the directories used by sDevTools.
}
