% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_snapshot_test.R
\name{build_snapshot_test}
\alias{build_snapshot_test}
\title{Build a snapshot test.}
\usage{
build_snapshot_test(
  fn_name,
  code,
  overwrite = FALSE,
  commit_git = TRUE,
  push_github = TRUE
)
}
\arguments{
\item{fn_name}{\code{[string]}}

\item{overwrite}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}

\item{commit_git}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}

\item{push_github}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}
}
\value{
\code{build_snapshot_test}: invisible(NULL) Writes a test file to the tests/thatthat directory
}
\description{
Build a snapshot test.
}
