% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTest.R
\name{build_test}
\alias{build_test}
\title{Auto build R code tests.}
\usage{
build_test(test_name, init, test_code, test_file = NULL, overwrite = F)
}
\arguments{
\item{test_name}{[character]  Must have an exact length of 1.}

\item{init}{[call]  Initial code to run before testing. Must be a bracketed expr.  NULL is ok.}

\item{test_code}{[call]   Must be a bracketed expr cotaining code to test in future.  Tests will be built based on the output of any expr that doesn't have an assignment. Usually this should be the function that the test is being built for.  All future tests will compare to the output of the code when build_test() was called.}

\item{test_file}{[character] Should the test file name differ from the test name?   Defaults to NULL.  If NULL, the test_name will be used as the file name.  When build_test(), 2 files will be created.  One in the tests/testthat/ dir and one in the test_source_files directory.  If a test fails, the original code will be located in the test_source_files for inspection.}

\item{overwrite}{[logical] Overwrite esixting test? Must have an exact length of 1.  Defaults to FALSE}
}
\value{
\code{build_test}: [NULL]
}
\description{
Auto build R code tests.
}
\examples{
build_test("sum",
init=NULL,
test_code={
sum(1:5)
sum(4:9)
sum(40:23)
})
}
