% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_test_from_examples.R
\name{build_test_from_examples}
\alias{build_test_from_examples}
\title{Build a test from existing function examples.}
\usage{
build_test_from_examples(fn_name, package = current_pkg())
}
\arguments{
\item{fn_name}{a function name}

\item{package}{[subset]  Possible values: c('installed_packages').  Defaults to current_pkg()}
}
\value{
\code{build_test_from_examples}: NULL
}
\description{
Build a test from existing function examples.
}
