% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bump_pkg_version.R
\name{bump_pkg_version}
\alias{bump_pkg_version}
\title{Bump the package's version.}
\usage{
bump_pkg_version(path = getwd())
}
\arguments{
\item{path}{\code{[character]}  Defaults to \code{getwd()}.}
}
\value{
\code{bump_pkg_version}: \code{[invisible(character)]} returns the version bump message
}
\description{
Bump the package's version.
}
