% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_git.R
\name{check_git}
\alias{check_git}
\title{Check if a directory is using GIT.}
\usage{
check_git(dir = getwd())
}
\arguments{
\item{dir}{\code{[string]}  Defaults to \code{getwd()}}
}
\value{
\code{check_git}: TRUE if using git, character message if not using git
}
\description{
Check if a directory is using GIT.
}
\examples{
 check_git(getwd())
}
