% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_isolated.R
\name{check_isolated}
\alias{check_isolated}
\title{Check code in an isolated R session.}
\usage{
check_isolated(test, package = current_pkg())
}
\arguments{
\item{test}{[call]}

\item{package}{[choice]  NULL is ok.  Defaults to current_pkg()}
}
\value{
\code{run_isolated}: TRUE if the run generated no error, else an error message of class character.
}
\description{
Check code in an isolated R session.
}
\examples{
 x <- 1
 run_isolated({
 x + 1
 })
 run_isolated({
 x <- 1
 x + 1
 })
}
