% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{Argument check for a package name.}
\usage{
check_package(package, null.ok = FALSE)
}
\arguments{
\item{package}{[string]}

\item{null.ok}{[logical]  Defaults to FALSE}
}
\value{
\code{check_package}: TRUE is package exists, Message if non-existent
}
\description{
Argument check for a package name.
}
\examples{
 check_package(package = 'data.table')
 check_package(package = 'rlang')
 check_package(package = 'a_non_existing_package')
}
