% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctionUsage.R
\name{checkFunctionUsage}
\alias{checkFunctionUsage}
\title{Check a function's usage across an environment.}
\usage{
checkFunctionUsage(name, package = current_pkg(), env = NULL)
}
\arguments{
\item{name}{\code{[string]}}

\item{package}{\code{[subset]}  Possible values: \code{installed_packages()}.  Defaults to \code{current_pkg()}}

\item{env}{\code{[environment]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{checkFunctionUsage}: data.table(obj_name,type,symbol_match,string_match)
}
\description{
Check a function's usage across an environment.
}
\examples{
 checkFunctionUsage('expr', package = 'sDevTools')
}
