% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackageUsage.R
\name{checkPackageUsage}
\alias{checkPackageUsage}
\title{Check the usage of all function internals in a package in an isolated environment.}
\usage{
checkPackageUsage(package = NULL)
}
\arguments{
\item{package}{[choice]  NULL is ok.  Defaults to NULL}
}
\value{
\code{checkPackageFns}: TRUE if no issues were found, otherwise a table of potential issues.
}
\description{
Check the usage of all function internals in a package in an isolated environment.
}
