% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUsageFn.R
\name{checkUsageFn}
\alias{checkUsageFn}
\title{Check the usage of fn internals in an isolated package environment.}
\usage{
checkUsageFn(fn, fn_name = NULL, package = NULL)
}
\arguments{
\item{fn}{[function]}

\item{package}{[choice]  NULL is ok.  Defaults to NULL}
}
\value{
\code{checkUsageFn}: TRUE if no issues were found, otherwise a table of potential issues.
}
\description{
Check the usage of fn internals in an isolated package environment.
}
\examples{
 fn <- function(x) {
 f(s, x)
 }
 checkUsageFn(fn, 'sDevTools')
}
