% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_all_to_github.R
\name{commit_all_to_github}
\alias{commit_all_to_github}
\title{Add, commit, and push all changes to github.}
\usage{
commit_all_to_github(path = getwd())
}
\arguments{
\item{path}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{commit_all_to_github}: \code{[invisible(NULL)]}
}
\description{
Add, commit, and push all changes to github.
}
