% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_github_repo.R
\name{create_github_repo}
\alias{create_github_repo}
\title{Initializes git / github repo for the current project.}
\usage{
create_github_repo(git_pat = git_token())
}
\value{
\code{create_github_repo}: [NULL]
}
\description{
Initializes git / github repo for the current project.
}
