% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentCursor.R
\name{currentCursor}
\alias{currentCursor}
\title{Get the current cursor position in the Rstudio editor.}
\usage{
currentCursor()
}
\value{
\code{currentCursor}: [named list('row.start','col.start','row.end','col.start','text')]
}
\description{
Get the current cursor position in the Rstudio editor.
}
