% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_description.R
\name{fill_description}
\alias{fill_description}
\title{Create a description file for an R package.}
\usage{
fill_description(
  pkg_name = "PKG_NAME",
  pkg_title = "PKG_TITLE",
  pkg_description = "Package Description.",
  author_first_name = "Theo",
  author_last_name = "Schrock",
  author_email = "theorschrock@gmail.com",
  path = getwd()
)
}
\arguments{
\item{pkg_name}{\code{[string]}  Defaults to \code{'PKG_NAME'}}

\item{pkg_title}{\code{[string]}  Defaults to \code{'PKG_TITLE'}}

\item{pkg_description}{\code{[string]}  Defaults to \code{'Package Description.'}}

\item{author_first_name}{\code{[string]}  Defaults to \code{'Theo'}}

\item{author_last_name}{\code{[string]}  Defaults to \code{'Schrock'}}

\item{author_email}{\code{[string]}  Defaults to \code{'<theorschrock@gmail.com>'}}

\item{path}{\code{[file]}  Defaults to \code{getwd()}}
}
\value{
\code{fill_description}: [NULL]
}
\description{
Create a description file for an R package.
}
