% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_document.R
\name{fn_document}
\alias{fn_document}
\title{Create function documentation for R packages.}
\usage{
fn_document(
  fn,
  examples = NULL,
  rdname = NULL,
  open = FALSE,
  overwrite = FALSE,
  package = current_pkg()
)
}
\arguments{
\item{fn}{[function]}

\item{examples}{[call]  NULL is ok.  Defaults to NULL}

\item{rdname}{[string]  NULL is ok.  Defaults to NULL}

\item{open}{[logical]  Defaults to TRUE}

\item{overwrite}{[logical]  Defaults to FALSE}

\item{package}{[subset]  Possible values: c('installed_packages').  Defaults to current_pkg()}
}
\value{
\code{fn_document}: [invisible(path)]
}
\description{
Create function documentation for R packages.
}
