% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_danger.R
\name{g_danger}
\alias{g_danger}
\title{Print a warning message to the console.}
\usage{
g_danger(message, env = caller_env())
}
\arguments{
\item{message}{\code{[string]}}

\item{env}{\code{[environment]}  Defaults to \code{caller_env()}}
}
\value{
\code{g_danger}: [print]
}
\description{
Print a warning message to the console.
}
\examples{
 type = 'dangerous'
 g_danger('this is a {type} message')
}
