% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_info.R
\name{g_info}
\alias{g_info}
\title{Print a warning message to the console.}
\usage{
g_info(message, env = caller_env())
}
\arguments{
\item{message}{\code{[string]}}

\item{env}{\code{[environment]}  Defaults to \code{caller_env()}}
}
\value{
\code{g_info}: [print]
}
\description{
Print a warning message to the console.
}
\examples{
 type = 'informative'
 g_info('this is a {type} message')
}
