% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_success.R
\name{g_success}
\alias{g_success}
\title{Print a success message to the console.}
\usage{
g_success(message, env = caller_env())
}
\arguments{
\item{message}{\code{[string]}}

\item{env}{\code{[environment]}  Defaults to \code{caller_env()}}
}
\value{
\code{g_success}: [NULL]
}
\description{
Print a success message to the console.
}
\examples{
 g_success('Success')
}
