% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_warn.R
\name{g_warn}
\alias{g_warn}
\title{Print a warning message to the console.}
\usage{
g_warn(message, env = caller_env())
}
\arguments{
\item{message}{\code{[string]}}

\item{env}{\code{[environment]}  Defaults to \code{caller_env()}}
}
\value{
\code{g_warn}: [print]
}
\description{
Print a warning message to the console.
}
\examples{
 type = 'warning'
 g_warn('this is a {type}')
}
