% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_imports_file.R
\name{generate_imports_file}
\alias{generate_imports_file}
\title{Generate imports.}
\usage{
generate_imports_file()
}
\value{
\code{generate_imports_file}: [NULL]
}
\description{
R file for package imports
}
