% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fn_examples.R
\name{get_fn_examples}
\alias{get_fn_examples}
\title{Get function examples from documentation.}
\usage{
get_fn_examples(fn_name)
}
\arguments{
\item{fn_name}{[file_exists]}
}
\value{
\code{get_fn_examples}: [expr(examples)]
}
\description{
Get function examples from documentation.
}
