% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_testing_deps.R
\name{get_project_testing_deps}
\alias{get_project_testing_deps}
\title{Retieve packages that should be load before running test.}
\usage{
get_project_testing_deps(dir = getwd())
}
\arguments{
\item{dir}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{get_project_testing_deps}: [character]
}
\description{
Retieve packages that should be load before running test.
}
