% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_names.R
\name{get_test_names}
\alias{get_test_names}
\title{Get test names in a dev package.}
\usage{
get_test_names(path = getwd())
}
\arguments{
\item{path}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{get_test_names}: \code{[character]}
}
\description{
Get test names in a dev package.
}
