% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjClasses.R
\name{getObjClasses}
\alias{getObjClasses}
\title{Get the classes of all objects in a package.}
\usage{
getObjClasses(env = package_env(), classtype = NULL)
}
\arguments{
\item{env}{\code{[Environment]} an environment}

\item{classtype}{\code{[character]} a vector of class names}
}
\value{
\code{getPackageObjClasses}: [named character]
}
\description{
Get the classes of all objects in a package.
}
