% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pkg.R
\name{import_pkg}
\alias{import_pkg}
\title{Add imports to the current dev package.}
\usage{
import_pkg(pkg_names)
}
\arguments{
\item{pkg_names}{[subset]  Possible values: c('installed_packages').}
}
\value{
\code{import_pkg}: NULL
}
\description{
Add imports to the current dev package.
}
