% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeTestthat.R
\name{initializeTestthat}
\alias{initializeTestthat}
\title{Initialize testthat and build test directories in the working directory.}
\usage{
initializeTestthat(test_deps = NULL)
}
\arguments{
\item{test_deps}{\code{[package]} Any packages that should be loaded before tests are ran. If project is a package, the package name in the working directory is automatically loaded.  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{initializeTestthat}: invisible(NULL)
}
\description{
Initialize testthat and build test directories in the working directory.
}
