% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installPrivate.R
\name{installPrivate}
\alias{installPrivate}
\title{Install a package theorschrock1 github.}
\usage{
installPrivate(package)
}
\arguments{
\item{package}{\code{[string]} the package name}
}
\value{
\code{installPrivate}: [NULL]
}
\description{
Install a package theorschrock1 github.
}
