% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dir_using_git.R
\name{is_dir_using_git}
\alias{is_dir_using_git}
\title{Check if a directory is using GIT.}
\usage{
is_dir_using_git(dir = getwd())
}
\arguments{
\item{dir}{\code{[string]}  Defaults to \code{getwd()}}
}
\value{
\code{is_dir_using_git}: [Logical(1)]
}
\description{
Check if a directory is using GIT.
}
\examples{
 is_dir_using_git(getwd())
}
