% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_internet_connected.R
\name{is_internet_connected()}
\alias{is_internet_connected()}
\alias{is_internet_connected}
\title{Check if system is connected to the internet}
\usage{
is_internet_connected()
}
\value{
\code{[logical(1)]}
}
\description{
Check if system is connected to the internet
}
\examples{
 is_internet_connected()
}
