% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_testthat_initialized.R
\name{is_testthat_initialized}
\alias{is_testthat_initialized}
\title{Test if test that is initialized in a dir.}
\usage{
is_testthat_initialized(dir = getwd())
}
\arguments{
\item{dir}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{is_testthat_initialized}: Logical(1)
}
\description{
Test if test that is initialized in a dir.
}
