% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch a shiny app package.}
\usage{
launch_app(package = current_pkg())
}
\arguments{
\item{package}{\code{[package]}  Defaults to \code{current_pkg()}}
}
\value{
\code{launch_app}: \code{[invisible(NULL)]}
}
\description{
Launch a shiny app package.
}
