% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load a vector of package names.}
\usage{
load_packages(pkgs)
}
\arguments{
\item{pkgs}{[subset]  Possible values: c('installed_packages').}
}
\value{
\code{load_packages}: [invisible(NULL)]
}
\description{
Load a vector of package names.
}
