% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDependencies.R
\name{loadDependencies}
\alias{loadDependencies}
\title{Load all package dependencies.}
\usage{
loadDependencies(package = current_pkg())
}
\arguments{
\item{package}{[character]  Defaults to current_pkg()}
}
\value{
\code{loadDependencies}: [NULL]
}
\description{
Load all package dependencies.
}
