% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadUtils.R
\name{loadUtils}
\alias{loadUtils}
\title{Load any file ending with utils.R into the global environment.}
\usage{
loadUtils()
}
\value{
\code{loadUtils}: NULL
}
\description{
Load any file ending with utils.R into the global environment.
}
