% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_packages.R
\name{local_packages}
\alias{local_packages}
\title{Get names of local R packages.}
\usage{
local_packages(root_dir = "~/")
}
\arguments{
\item{root_dir}{\code{[character]}  Defaults to \code{'~/'}}
}
\value{
\code{local_packages}: \code{[character]}
}
\description{
Get names of local R packages.
}
