% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fn_asssertion.R
\name{make_fn_asssertion}
\alias{make_fn_asssertion}
\title{Make assertion for a new function.}
\usage{
make_fn_asssertion(dots)
}
\arguments{
\item{dots}{\code{[call]}  Must be a call from function \code{'.'}.}
}
\value{
\code{make_fn_asssertion}: \code{[list(exprs)]}
}
\description{
Make assertion for a new function.
}
\examples{
 x = expr(.(id = string(), name = string(), env = env(caller_env),
 args = named_list(structure = list(x = num(), y = num()))))
 make_fn_asssertion(x)
}
