% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dev_function.R
\name{new_dev_function}
\alias{new_dev_function}
\title{Create a new dev function template.}
\usage{
new_dev_function(fn_name, ..., type = "standard", package = current_pkg())
}
\arguments{
\item{fn_name}{[string]}

\item{type}{[choice('standard'|'shiny')] function template type}

\item{package}{[subset]  Possible values: c('installed_packages').  Defaults to current_pkg()}
}
\value{
\code{new_dev_function}: [NULL]
}
\description{
Create a new dev function template.
}
\examples{
 new_dev_function('newfn')
}
