% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_fn.R
\name{new_fn}
\alias{new_fn}
\title{Create a package function template.}
\usage{
new_fn(fn_args, desc, return, type = "standard")
}
\arguments{
\item{fn_args}{\code{[call]} \code{function_name(...)}}

\item{desc}{\code{[string]} what the function does.}

\item{return}{\code{[string]} what the function returns.}

\item{type}{\code{[choice]}  Possible values: \code{c('standard', 'shiny')}. If shiny, the examples will auto populate with a shiny app template. Defaults to 'standard'.}
}
\value{
\code{new_fn}:\code{[NULL]}
}
\description{
Create a package function template.
}
\examples{
 if (interactive()) {
 new_fn(
    myFunc(
      x = numeric(lower = 0, upper = 1),
      y = numeric(lower = 0, upper = 1),
        na.rm = TF(TRUE)
        ),
 desc = 'What this function does',
 return = 'What this function returns')
 }
}
