% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mod.R
\name{new_mod}
\alias{new_mod}
\title{Create a shiny module template.}
\usage{
new_mod(fn_args, desc, return)
}
\arguments{
\item{fn_args}{\code{[call]} \code{function_name(...)}}

\item{desc}{\code{[string]} what the function does.}

\item{return}{\code{[string]} what the function returns.}
}
\value{
\code{new_fn}:\code{[NULL]}
}
\description{
Create a shiny module template.
}
\examples{
 if (interactive()) {
 new_mod(
    myMod(
      id=string(),
       x=reactive(type='numeric',lower = 0, upper = 1),
       y=reactive(type='numeric',lower = 0, upper = 1),
        na.rm =reactive(type='logical',len=1)
        ),
 desc = 'What this module does',
 return = 'What this module returns')
 }
}
