% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_shiny_input_binding.R
\name{new_shiny_input_binding}
\alias{new_shiny_input_binding}
\title{Create a new shiny input binding.}
\usage{
new_shiny_input_binding(name)
}
\value{
\code{new_shiny_input_binding}: \code{[invisible(NULL)]}
}
\description{
Create a new shiny input binding.
}
