% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newShortCutBinding.R
\name{newShortCutBinding}
\alias{newShortCutBinding}
\title{Create a custom key board binding.}
\usage{
newShortCutBinding(
  fn,
  name,
  description,
  package = current_pkg(),
  interactive = TRUE
)
}
\arguments{
\item{fn}{[function]}

\item{name}{[string]}

\item{description}{[string]}

\item{package}{[string]  Defaults to current_pkg()}

\item{interactive}{[logical]  Must have an exact length of 1.  Defaults to TRUE}
}
\value{
\code{newShortCutBinding}: NULL
}
\description{
Create a custom key board binding.
}
\examples{
 if (interactive()) {
 library(praise)
 newShortCutBinding(praise, name = 'Say something nice',
 description = 'Print a compliment to the console',
 package = 'praise', interactive = TRUE)
 }
}
