% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_detect_name.R
\name{obj_detect_name}
\alias{obj_detect_name}
\title{Detect a name reference in an function or object.}
\usage{
obj_detect_name(obj, name)
}
\arguments{
\item{obj}{an R object}

\item{name}{\code{[string]}}
}
\value{
\code{obj_detect_name}: data.table(obj_name,type,symbol_match,string_match)
}
\description{
Detect a name reference in an function or object.
}
\examples{
 fn = function(x) {
 map(x, function(x) glue('the sum is {sum(x)}'))
 }
 obj_detect_name(fn, name = 'map')
 obj_detect_name(fn, name = 'sum')
}
