% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onPackageLoad.R
\name{onPackageLoad}
\alias{onPackageLoad}
\title{Add functionality to the package's onLoad process.}
\usage{
onPackageLoad(code, append = TRUE, message = NULL, package = getwd())
}
\arguments{
\item{append}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}

\item{message}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}

\item{package}{\code{[sdevtools_proj]}  Defaults to \code{getwd()}}
}
\value{
\code{onPackageLoad}: invisible(NULL) Writes to the .onLoad function in the imports.R file.
}
\description{
Add functionality to the package's onLoad process.
}
