% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fn_source.R
\name{open_fn_source}
\alias{open_fn_source}
\title{Open a function source file.}
\usage{
open_fn_source(fn_name)
}
\arguments{
\item{fn_name}{\code{[string]}}
}
\value{
\code{open_fn_source}: invisible(null)
}
\description{
Open a function source file.
}
