% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_guide.R
\name{open_guide}
\alias{open_guide}
\title{Open a 'how to' guide.}
\usage{
open_guide(guide)
}
\arguments{
\item{guide}{[file_exists]}
}
\value{
\code{open_guide}: [NULL]
}
\description{
Open a 'how to' guide.
}
