% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package2list.R
\name{package2list}
\alias{package2list}
\title{list functions in a dev package.}
\usage{
package2list(path = getwd())
}
\arguments{
\item{path}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{package2list}: [list(exprs)]
}
\description{
list functions in a dev package.
}
