% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_env.R
\name{package_env}
\alias{package_env}
\title{Get a packages enviromnent.}
\usage{
package_env(package = current_pkg())
}
\arguments{
\item{package}{\code{[subset]}  Possible values: \code{installed_packages()}.  Defaults to \code{current_pkg()}}
}
\value{
\code{package_env}: [Environment]
}
\description{
Get a packages enviromnent.
}
