% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_file.R
\name{package_file}
\alias{package_file}
\title{create a file path to an install R package.}
\usage{
package_file(package, path)
}
\arguments{
\item{package}{\code{[string]}}

\item{path}{\code{[string]}}
}
\value{
\code{package_file}: [character(file_path)]
}
\description{
create a file path to an install R package.
}
