% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_fn_names.R
\name{package_fn_names}
\alias{package_fn_names}
\title{Get object names in a dev package.}
\usage{
package_fn_names(path = getwd())
}
\arguments{
\item{path}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{package_fn_names}: [character]
}
\description{
Get object names in a dev package.
}
