% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_test_names.R
\name{package_test_names}
\alias{package_test_names}
\title{Get test names in a dev package.}
\usage{
package_test_names(path = getwd())
}
\arguments{
\item{path}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{package_test_names}: [character]
}
\description{
Get test names in a dev package.
}
