% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_file.R
\name{parse_file}
\alias{parse_file}
\title{Parse an r file.}
\usage{
parse_file(path)
}
\arguments{
\item{path}{[file_exists]}
}
\value{
\code{parse_file}: [list(exprs)] unevalualted R expression
}
\description{
Parse an r file.
}
