% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_deps.R
\name{pkg_deps}
\alias{pkg_deps}
\title{Get package dependencies.}
\usage{
pkg_deps(package = current_pkg())
}
\arguments{
\item{package}{[subset]  Possible values: c('installed_packages').  Defaults to current_pkg()}
}
\value{
\code{pkg_deps}: [character]
}
\description{
Get package dependencies.
}
