% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_start_up.R
\name{pkg_start_up}
\alias{pkg_start_up}
\title{Generate a R Package Start-up file.}
\usage{
pkg_start_up(pkg_name)
}
\arguments{
\item{pkg_name}{[string]}
}
\value{
\code{pkg_start_up}: NULL
}
\description{
Generate a R Package Start-up file.
}
