% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6active.R
\name{r6active}
\alias{r6active}
\title{insert an R6 active template into the editor.}
\usage{
r6active(name, read_only = T)
}
\arguments{
\item{name}{the R6 class name}

\item{read_only}{\code{[logical]}  Defaults to \code{T}}
}
\value{
\code{r6active}: NULL
}
\description{
insert an R6 active template into the editor.
}
