% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6skeleton.R
\name{r6skeleton}
\alias{r6skeleton}
\title{insert an R6 skeleton into the editor.}
\usage{
r6skeleton(name = "TmpName", inherits = NULL)
}
\arguments{
\item{name}{\code{[string]}  Defaults to \code{'TmpName'}}

\item{inherits}{\code{[string]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{r6skeleton}: NULL
}
\description{
insert an R6 skeleton into the editor.
}
