% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_test.R
\name{rebuild_test}
\alias{rebuild_test}
\title{Get a test to rebuid.}
\usage{
rebuild_test(testname)
}
\value{
\code{rebuild_test}: invisible(NULL) Insert the current test at the cursor
}
\description{
Get a test to rebuid.
}
