% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_dev.R
\name{refresh_dev}
\alias{refresh_dev}
\title{Clear the current dev file.}
\usage{
refresh_dev(package = current_pkg())
}
\arguments{
\item{package}{[subset]  Possible values: c('installed_packages').  Defaults to current_pkg()}
}
\value{
\code{refresh_dev}: [NULL]
}
\description{
Clear the current dev file.
}
\examples{
 refresh_dev()
}
