% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_import.R
\name{remove_import}
\alias{remove_import}
\title{Remove imports to the current dev package.}
\usage{
remove_import(pkg_names)
}
\arguments{
\item{pkg_names}{[character]}
}
\value{
\code{remove_import}: NULL
}
\description{
Remove imports to the current dev package.
}
