% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restartAndReload.R
\name{restartAndReload}
\alias{restartAndReload}
\title{Restart and reload environment.}
\usage{
restartAndReload()
}
\value{
\code{restartAndReload}: NULL
}
\description{
Restart and reload environment.
}
