% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_comments.R
\name{rox_comments}
\alias{rox_comments}
\title{Create roxygen comments from a string.}
\usage{
rox_comments(x)
}
\arguments{
\item{x}{\code{[character]}}
}
\value{
\code{rox_comments}: [character]
}
\description{
Create roxygen comments from a string.
}
\examples{
 rox_comments('these are roxygen comments\n  more comments\n  more comments\n  @noRd')
}
