% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isolated.R
\name{run_isolated}
\alias{run_isolated}
\title{Run R code in an isolated R session.}
\usage{
run_isolated(test, package = current_pkg())
}
\arguments{
\item{test}{[call]}

\item{package}{[choice]  NULL is ok.  Defaults to current_pkg()}
}
\value{
\code{run_isolated}: list of results
}
\description{
Run R code in an isolated R session.
}
\examples{
 x <- 1
 run_isolated({
 x + 1
 })
 run_isolated({
 x <- 1
 x + 1
 })
}
