% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{runTests}
\alias{runTests}
\title{Run all tests in the testthat dir of a package.}
\usage{
runTests(name = NULL, package = getwd(), dev_version = FALSE)
}
\arguments{
\item{name}{\code{[string]} the test name}

\item{package}{\code{[string]}  Defaults to current_pkg()}

\item{dev_version}{\code{[logical(1)]} If true, the test will be run using \code{devtools::load_all()} rather than \code{library(PKG_NAME)}.}
}
\value{
\code{runTests}: [Print(TestResults)]
}
\description{
Run all tests in the testthat dir of a package.
}
