% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sDevTools_initialize.R
\name{sDevTools_initialize}
\alias{sDevTools_initialize}
\title{Generate a sDevTools start up R script.}
\usage{
sDevTools_initialize(package_name)
}
\arguments{
\item{package_name}{\code{[string]}}
}
\value{
\code{sDevTools_initialize}: [Character]
}
\description{
Generate a sDevTools start up R script.
}
