% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchActiveEditor.R
\name{searchActiveEditor}
\alias{searchActiveEditor}
\title{Search the active Rstudio editor for regex patterns.}
\usage{
searchActiveEditor(pattern)
}
\arguments{
\item{pattern}{[string]}
}
\value{
\code{searchActiveEditor}: [data.table]  rowid, start, and end for all matches
}
\description{
Search the active Rstudio editor for regex patterns.
}
