% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_example_template.R
\name{shiny_example_template}
\alias{shiny_example_template}
\title{Generate a shiny app template.}
\usage{
shiny_example_template(name, args)
}
\arguments{
\item{name}{the name of the primary input}
}
\value{
\code{shiny_example_template}: a shiny app template
}
\description{
Generate a shiny app template.
}
\examples{
 shiny_example_template('new_template')
}
