% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressUsageWarnings.R
\name{suppressUsageWarnings}
\alias{suppressUsageWarnings}
\title{Suppress matching future warnings generated by checkPackageFns.}
\usage{
suppressUsageWarnings(warnings, package = NULL)
}
\arguments{
\item{warnings}{[class]}

\item{package}{[choice]  NULL is ok.  Defaults to NULL}
}
\value{
\code{suppressUsageWarnings}: NULL
}
\description{
Suppress matching future warnings generated by checkPackageFns.
}
