% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_named_list.R
\name{test_named_list}
\alias{test_named_list}
\title{test.}
\usage{
test_named_list(x, y)
}
\arguments{
\item{x}{\code{[named_list]}  The list must have the following structure: \rcode{list(x = char(), y = string())}.}

\item{y}{\code{[reactive]}  Must be a reactive expression which has an output type: \rcode{['named_list']}.  The list must have the following structure: \rcode{list(x = char(), y = string())}.}
}
\value{
\code{test_named_list}: test
}
\description{
test.
}
