% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untested_package_fns.R
\name{untested_package_fns}
\alias{untested_package_fns}
\title{Get names of untested function in a dev package.}
\usage{
untested_package_fns(path = getwd())
}
\arguments{
\item{path}{\code{[directory]}  Defaults to \code{getwd()}}
}
\value{
\code{untested_package_fns}: [character]
}
\description{
Get names of untested function in a dev package.
}
