% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gitpat_renviron.R
\name{use_gitpat_renviron}
\alias{use_gitpat_renviron}
\title{Creates an Renviron file with GITHUB_PAT}
\usage{
use_gitpat_renviron(git_pat = git_token())
}
\value{
\code{use_gitpat_renviron}: NULL
}
\description{
Creates an Renviron file with GITHUB_PAT
}
