% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_snapshot.R
\name{wrap_spanshot}
\alias{wrap_spanshot}
\alias{wrap_snapshot}
\title{Any line without an assignment will be wrapped in \code{expect_snapshot()}.}
\usage{
wrap_snapshot(lines)
}
\arguments{
\item{lines}{\code{[exprs]} the expressions to wrap.  any line without an assignment  that should be ignore should be wrapped in  \code{ignore()}}
}
\value{
\code{[exprs]}
}
\description{
Any line without an assignment will be wrapped in \code{expect_snapshot()}.
}
