% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadUtils.R
\name{loadUtils}
\alias{loadUtils}
\title{Load all non-exported functions into the package environment}
\usage{
loadUtils()
}
\value{
\code{loadUtils}: NULL
}
\description{
Load all non-exported functions into the package environment
}
