% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{runTests}
\alias{runTests}
\title{Run all tests in the testthat dir of a package.}
\usage{
runTests(name = NULL, package = current_pkg())
}
\arguments{
\item{name}{[string] the test name}

\item{package}{[string]  Defaults to current_pkg()}
}
\value{
\code{runTests}: [Print(TestResults)]
}
\description{
Run all tests in the testthat dir of a package.
}
